+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `PMG\Queue\Driver::enqueue` now typehints agains `object` instead of message.
   if a driver gets an `Envelope` instance it should use that instead of creating
   its own envelope. See [`UPGRADE-5.0.md`](/UPGRADE-5.0.md) for more details.
+- `PMG\Queue\MessageHandler::handle` now typehints against `object` instead of
+  `Message`.
 - `MessageLifecycle::failed` no longer has an `$isRetrying` argument, instead
   `MessageLifecycyle::retrying` will be called instead.
 - The `Message` interface no longer has a `getName` method.